+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/RawMessage.kt
@@ -148,7 +148,7 @@ internal data class RawMessage(
             )
             forward_from_chat is ChannelChat -> ForwardFromChannelInfo(
                 forward_date,
-                forward_from_message_id ?: throw IllegalStateException("Channel forwarded message must contain message id, but was not"),
+                forward_from_message_id ?: error("Channel forwarded message must contain message id, but was not"),
                 forward_from_chat,
                 forward_signature
             )
@@ -209,7 +209,7 @@ internal data class RawMessage(
                         chatEvent as? ChannelEvent ?: throwWrongChatEvent(ChannelEvent::class, chatEvent),
                         date.asDate
                     )
-                    else -> throw IllegalStateException("Expected one of the public chats, but was $chat (in extracting of chat event message)")
+                    else -> error("Expected one of the public chats, but was $chat (in extracting of chat event message)")
                 }
             } ?: content?.let { content ->
                 media_group_id?.let {
@@ -222,7 +222,7 @@ internal data class RawMessage(
                             when (content) {
                                 is PhotoContent -> content
                                 is VideoContent -> content
-                                else -> throw IllegalStateException("Unsupported content for media group")
+                                else -> error("Unsupported content for media group")
                             },
                             edit_date?.asDate,
                             forwarded,
@@ -238,7 +238,7 @@ internal data class RawMessage(
                             when (content) {
                                 is PhotoContent -> content
                                 is VideoContent -> content
-                                else -> throw IllegalStateException("Unsupported content for media group")
+                                else -> error("Unsupported content for media group")
                             },
                             edit_date?.asDate,
                             forwarded,
@@ -261,8 +261,7 @@ internal data class RawMessage(
                     )
                     else -> CommonMessageImpl(
                         messageId,
-                        from
-                            ?: throw IllegalStateException("Was detected common message, but owner (sender) of the message was not found"),
+                        from ?: error("Was detected common message, but owner (sender) of the message was not found"),
                         chat,
                         content,
                         date.asDate,
@@ -274,7 +273,7 @@ internal data class RawMessage(
                         paymentInfo
                     )
                 }
-            } ?: throw IllegalStateException("Was not found supported type of data")
+            } ?: error("Was not found supported type of data")
         } catch (e: Exception) {
             UnknownMessageType(
                 messageId,
@@ -286,6 +285,6 @@ internal data class RawMessage(
     }
 
     private fun throwWrongChatEvent(expected: KClass<*>, but: ChatEvent): CommonEvent {
-        throw IllegalStateException("Wrong type of chat event: expected $expected, but was $but")
+        error("Wrong type of chat event: expected $expected, but was $but")
     }
 }